Move AssetResult to new file. Refactor all assets processor methods to use processedAssetName for name validation. Code style cleanup. Cleanup asset processing variable names to refer to assets instead of colors.
